Enable specifying the Endpoint Suffix in the azure connection string
The EndpointSuffix of an azure storage connection string changes depending on the Azure environment. The suffix for public Azure clouds is core.windows.net (default). The suffix for Azure USGov cloud is core.usgovcloudapi.net. This change allows the client to use the EndpointSuffix specified in the azure configuration file.